#1861 Moved all gokernargument named functions into a separate file and cleaned up minor coding issues.
