+++ src/psyclone/tests/gocean1p0_test.py
@@ -39,26 +39,24 @@
 '''Tests for PSy-layer code generation that are specific to the
 GOcean 1.0 API.'''
 
-from __future__ import absolute_import, print_function
 import os
 import re
+
 import pytest
+
 from psyclone.configuration import Config
-from psyclone.parse.algorithm import parse, Arg
-from psyclone.parse.kernel import Descriptor
+from psyclone.parse.algorithm import parse
 from psyclone.parse.utils import ParseError
 from psyclone.errors import InternalError, GenerationError
 from psyclone.psyGen import PSyFactory
-from psyclone.gocean1p0 import GOKern, GOLoop, \
-    GOKernelArgument, GOKernelArguments, GOKernelGridArgument, \
-    GOBuiltInCallFactory, GOSymbolTable
+from psyclone.gocean1p0 import (GOKern, GOLoop, GOKernelArgument,
+                                GOKernelGridArgument, GOBuiltInCallFactory,
+                                GOSymbolTable)
 from psyclone.tests.utilities import get_invoke
 from psyclone.tests.gocean_build import GOceanBuild
-from psyclone.psyir.symbols import SymbolTable, DeferredType, \
-    ContainerSymbol, DataSymbol, ImportInterface, ScalarType, INTEGER_TYPE, \
-    ArgumentInterface, DataTypeSymbol
-from psyclone.psyir.nodes import Node, StructureReference, Member, \
-    StructureMember, Reference, Literal
+from psyclone.psyir.symbols import (DeferredType, ContainerSymbol, DataSymbol,
+                                    ImportInterface, INTEGER_TYPE,
+                                    ArgumentInterface)
 from psyclone.domain.gocean.transformations import GOConstLoopBoundsTrans
 
 API = "gocean1.0"
@@ -1421,268 +1419,6 @@ def test14_no_builtins():
     with pytest.raises(GenerationError) as excinfo:
         GOBuiltInCallFactory.create()
     assert "Built-ins are not supported for the GOcean" in str(excinfo.value)
-
-
-def test_gokernelarguments_append():
-    ''' Check the GOcean specialisation of KernelArguments append method'''
-
-    # Parse a file to get an initialised GOKernelsArguments object
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    symtab = invoke.schedule.symbol_table
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-    assert isinstance(argument_list, GOKernelArguments)
-
-    # Try append a non-string value
-    with pytest.raises(TypeError) as err:
-        argument_list.append(3, "space")
-    assert "The name parameter given to GOKernelArguments.append method " \
-           "should be a string, but found 'int' instead." in str(err.value)
-
-    # Append well-constructed arguments
-    argument_list.append(symtab.new_symbol("var1").name, "go_r_scalar")
-    argument_list.append(symtab.new_symbol("var2").name, "go_i_scalar")
-
-    assert isinstance(kernelcall.args[-1], GOKernelArgument)
-    assert isinstance(kernelcall.args[-2], GOKernelArgument)
-    assert kernelcall.args[-1].name == "var2"
-    assert kernelcall.args[-2].name == "var1"
-
-    # And the generated code looks as expected
-    generated_code = str(psy.gen)
-    assert "CALL compute_cu_code(i, j, cu_fld%data, p_fld%data, u_fld%data," \
-           " var1, var2)" in generated_code
-
-
-def test_gokernelargument_infer_datatype():
-    ''' Check the GOcean specialisation of the infer_datatype works for each
-    possible type of KernelArgument. '''
-
-    # Parse an invoke with a scalar float and a field
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_scalar_float_arg.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-
-    # The first argument is a scalar Real
-    datatype = argument_list.args[0].infer_datatype()
-    assert isinstance(datatype, ScalarType)
-    assert datatype.intrinsic == ScalarType.Intrinsic.REAL
-    assert datatype.precision.name == "go_wp"
-
-    # The second argument is a r2d_field (imported DataTypeSymbol)
-    assert isinstance(argument_list.args[1].infer_datatype(), DataTypeSymbol)
-    assert argument_list.args[1].infer_datatype().name == "r2d_field"
-
-    # Parse an invoke with a scalar int and a field
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_scalar_int_arg.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-
-    # The first argument is a scalar Integer
-    assert argument_list.args[0].infer_datatype() == INTEGER_TYPE
-
-    # The second argument is a r2d_field (imported DataTypeSymbol)
-    assert isinstance(argument_list.args[1].infer_datatype(), DataTypeSymbol)
-    assert argument_list.args[1].infer_datatype().name == "r2d_field"
-
-    # Test an incompatible Kernel Argument
-    argument_list.args[0]._arg._space = "incompatible"
-    with pytest.raises(InternalError) as excinfo:
-        _ = argument_list.args[0].infer_datatype()
-    assert ("GOcean expects scalar arguments to be of 'go_r_scalar' or "
-            "'go_i_scalar' type but found 'incompatible'."
-            in str(excinfo.value))
-
-    argument_list.args[0]._arg._argument_type = "incompatible"
-    with pytest.raises(InternalError) as excinfo:
-        _ = argument_list.args[0].infer_datatype()
-    assert ("GOcean expects the Argument.argument_type() to be 'field' or "
-            "'scalar' but found 'incompatible'." in str(excinfo.value))
-
-
-def test_gokernelargument_intrinsic_type():
-    ''' Check that the GOcean specialisation of the intrinsic_type returns the
-    expected values. '''
-
-    # Parse an invoke with a scalar float and a field
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_scalar_float_arg.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments.args
-    # First argument 'a_scalar' is a REAL
-    assert argument_list[0].intrinsic_type == "real"
-    # Second argument 'ssh_fld' is a derived type and doesn't have a single
-    # intrinsic type, so it returns an empty string
-    assert argument_list[1].intrinsic_type == ""
-    # Change the first argument metadata type to integer, and check the
-    # intrinsic_type value also changes
-    argument_list[0]._arg._space = "go_i_scalar"
-    assert argument_list[0].intrinsic_type == "integer"
-
-
-def test_gokernelarguments_psyir_expressions():
-    ''' Check the GOcean specialisation of psyir_expressions returns the
-    expected list of PSyIR expressions for each argument'''
-
-    # Parse an invoke with grid properties
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_grid_props.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments.psyir_expressions()
-
-    # It has 2 indices arguments plus the kernel arguments
-    assert len(argument_list) == len(kernelcall.arguments.args) + 2
-
-    # Second argument is a reference to the symbol tagged contiguous_kidx
-    assert isinstance(argument_list[0], Reference)
-    assert (argument_list[0].symbol is
-            kernelcall.scope.symbol_table.lookup_with_tag("contiguous_kidx"))
-
-    # Second argument is a reference to the symbol tagged noncontiguous_kidx
-    assert isinstance(argument_list[1], Reference)
-    assert (argument_list[1].symbol is
-            kernelcall.scope.symbol_table.lookup_with_tag(
-                "noncontiguous_kidx"))
-
-    # Other arguments are also PSyIR expressions generated depending on the
-    # argument type. In this case it has 5 more arguments, all of them are
-    # structure references.
-    for argument in argument_list[2:7]:
-        assert isinstance(argument, StructureReference)
-
-
-def test_gokernelargument_psyir_expression():
-    ''' Check the GOcean specialisation of psyir_expression returns the
-    expected expression for any GOKernelArgument and GOKernelGridArguments'''
-
-    # Parse an invoke with grid properties
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_grid_props.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-
-    # The first argument is a field
-    expr1 = argument_list.args[0].psyir_expression()
-    assert isinstance(expr1, StructureReference)
-    assert isinstance(expr1.member, Member)
-    assert expr1.member.name == "data"
-
-    # Third argument is a tmask grid property
-    expr2 = argument_list.args[2].psyir_expression()
-    assert isinstance(expr2, StructureReference)
-    assert isinstance(expr2.member, StructureMember)
-    assert isinstance(expr2.member.member, Member)
-    assert expr2.member.name == "grid"
-    assert expr2.member.member.name == "tmask"
-
-    # Parse an invoke with a scalar int and a field
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_scalar_int_arg.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-
-    # The first argument is a scalar
-    expr3 = argument_list.args[0].psyir_expression()
-    assert isinstance(expr3, Reference)
-
-    # Test an incompatible Kernel Argument
-    argument_list.args[0]._arg._argument_type = "incompatible"
-    with pytest.raises(InternalError) as excinfo:
-        _ = argument_list.args[0].psyir_expression()
-    assert ("GOcean expects the Argument.argument_type() to be 'field' or "
-            "'scalar' but found 'incompatible'." in str(excinfo.value))
-
-
-def test_gokernelargument_constant_psyir_expression():
-    '''Test various constant arguments and their conversion to PSyIR.
-    '''
-
-    # Parse an invoke with a scalar int and a field
-    _, invoke_info = parse(os.path.join(os.path.
-                                        dirname(os.path.
-                                                abspath(__file__)),
-                                        "test_files", "gocean1p0",
-                                        "single_invoke_scalar_float_arg.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    invoke = psy.invokes.invoke_list[0]
-    kernelcall = invoke.schedule.coded_kernels()[0]
-    argument_list = kernelcall.arguments
-
-    for (const, intr_type) in [("1", ScalarType.Intrinsic.INTEGER),
-                               ("1.0", ScalarType.Intrinsic.REAL),
-                               ("1.0e+0", ScalarType.Intrinsic.REAL),
-                               ("1.0E-0", ScalarType.Intrinsic.REAL)]:
-        argument_list.args[0]._name = const
-        expr = argument_list.args[0].psyir_expression()
-        assert isinstance(expr, Literal)
-        assert expr.datatype.intrinsic == intr_type
-
-
-def test_gokernelargument_type(monkeypatch):
-    ''' Check the type property of the GOKernelArgument'''
-
-    # Create a dummy node with the symbol_table property
-    dummy_node = Node()
-    dummy_node.symbol_table = SymbolTable()
-
-    # Create a dummy GOKernelArgument
-    descriptor = Descriptor(None, "go_r_scalar")
-    arg = Arg("variable", "arg", "arg")
-    argument = GOKernelArgument(descriptor, arg, dummy_node)
-
-    # If the descriptor does not have a type it defaults to 'scalar'
-    assert argument.argument_type == "scalar"
-
-    # Otherwise it returns the descriptor type
-    # Mock the descriptor type method
-    monkeypatch.setattr(argument._arg, "_argument_type", "descriptor_type")
-    assert argument.argument_type == "descriptor_type"
 
 
 def test_gosymboltable_conformity_check():
